Fix :: CommonCode, WashInfo 에서 테이블과 데이터 타입이 맞지 않는 필드 수정

+++ module-api/src/test/java/com/kernel360/commoncode/controller/CommonCodeControllerTest.java
@@ -21,7 +21,7 @@ class CommonCodeControllerTest extends ControllerTest {
         String pathVariable = "color";
         Long codeNo = 1L;
         String codeName = "테스트용 코드";
-        Integer upperNo = null;
+        Long upperNo = null;
         String upperName = null;
         Integer sortOrder = 1;
         Boolean isUsed = true;
